Remove the need for a flag to prevent stealth address reuse
The primary goal of this commit is to prevent the need for an
extra flag, written to storage during Umbra sends, formely used to
prevent stealth address re-use.

In normal usage of the protocol, addresses would not be re-used, but
someone could call the contract methods with an address that had already
received funds. There is no downside to this, except in the case of token
payments, where the funds are held in the contract until withdrawn, and
the balance of each stealth address is held in the contract. In such a
case, re-using the address would cause the prior balance data to be
overwritten, effectively locking the funds.

The motivation for this change is to save gas. We want to avoid the extra
storage write without the risk of locked funds. To do this, we:

1. Allow re-use of addresses when the asset being sent is ETH. There is no
  downside to doing so and no risk of funds being locked, since they are
  simply forwarded to the stealth address
2. Use a non-zero current-token balance as a chekk to prevent address reuse
   in the specific case that the user has not yet withdrawn their tokens.
   This prevents funds from being locked, w/o writing extra data.