Add a generic method for signature based meta-tx withdrawal
Adds a new method to the core Umbra contract which enables
withdrawal from a stealth address with only a signature
from said stealth address. Stealth address holder must sign
a message designating the sponsor, i.e. the relayer, that
will be paid and the fee that will be paid to them out of the
token balance. Anyone can submit the tx, but only the relayer
has any incentive to do so.

The method does not use EIP-712 because it will be signed by
the *stealth* address key, not the key the user has in their
standard wallet. The stealth address key is generated by a
user signature from their wallet, and is subsequently in memory,
so the user does not need to approve the meta-tx signature.