Implement an UmbraRelayRecipient wrapper for GSN
We implement a new wrapper contract which extends OpenGSN's
BaseRelayRecipient and simply forwards GSN calls through to
the meta withdraw method on the base Umbra contract. This
is how Umbra can leverage GSN without having to support it at
the base contract level.

The tests now deploy an instance of UmbraRelayRecipient, then
make all GSN calls through this contract rather than on Umbra
directly. Note that we now must sign the messages directly
before passing to GSN, because our base withdraw method checks
the signature.

The GSN forwarder is also validating the signature from the sender,
which is now not neccessary. Until we write a custom forwarder which
ignores this step, we're buring additional gas on GSN calls.