fix: arbitrum eth withdrawals now work (reverts a3e9648)
The Nitro upgrade gave arbitrum the same gas metering as the EVM.
However, there is still an L1 gas cost. This aspect of pricing is baked
into the gas limit returned from eth_estimateGas. Therefore we should
not treat gas limits as a fixed value of 21k, and should continue to use
the same approach as before. More info on current Arbitrum gas behavior:
- https://developer.offchainlabs.com/arbos/gas
- https://developer.offchainlabs.com/arbos/l1-pricing
- https://medium.com/offchainlabs/understanding-arbitrum-2-dimensional-fees-fd1d582596c9